Major rework of the photohunter activity.
In order to make an level in this activity it was mandatory to
provide the differences between the 2 images in a text files.

In this commit, the differences are calculated by the computer which
makes it easier to create content.

I also added new content for this activity based on images found
in GCompris itself.