Created a new activity engine called findit. Replaced the color activity to use it.
This new activity makes it easy to create a new activity from a single data file.
The concept here is to provide a set of objects and the children will be proposed them.
The children have to click on the correct one.
It can be used to find an object by the image of it or by a voice, or both.