core: now configuration dialog callback requires a boolean return value indicating if the entered data are correct.
Now when we click ok in the config dialog. If the callback returns FALSE then it is not closed.
It is up to the dialog implementor to display a relevant error message.

I found out that the radio_buttons API makes the pythontest config dialog crashing.
Since it is not used elsewhere, I just commented it until we need it.

In login.py and tuxpaint, reactivated the config button that was missing.