click_on_letter: mostly full rewrite of this activity to allow a specific configuration per language and per user.
Now this activity takes a desktop formated file as input. It contains for each level the questions
and answers charaters to propose.

This file can be edited in the game through the config button, in this case a local
data file is save under the user's My GCompris folder and is reused.

Now for each country we must provide a data file, this is much better than before where
it was fully random and was not respecting the natural difficulties of a language.

Still miss a couple of feature:
- create again a random mode from the alphabet in case we don't have a dataset.
- in the dataset add a mode where the letter to find is displayed intead of spoken.