Fix crash with params x/y of wrong datatype.
The functionparameters x and y are sometimes float instead of int.
This leads to a crash at "xrange(y)" in the code below. This happens
for example with GIMP v2.6.12 and python v2.7.3 on a gentoo linux
system. Since x and y are always integer values - the spinners
configuration won't allow other values - we can safely convert to int
here.