Fix: mining-activity: no scrolling while bonus
We remember the pause state, so we can block input (scrolling /
mouseclicks), while in pause.
This forced to make the is_game_paused() function to be availably in the
Viewport class, so the Viewport's contructor had to be changed, too:
In order to pass every callback function, we pass the activity object
itself and the Viewport takes, what it needs.