Fix: mining-acitivity: no nugget behind gc-bar
There are certain places on the screen, where no gold nugget should be
placed. This commit introduces a mechanism to define those areas and to
test possible nugget places against those constraints.

So far: A nugget will not be placed behind the GCompris bar in the
lower, left corner and not in front of the wood pillars on the screen
edges.