Add different difficulties to mining activity
The maximum zoom depth and the number of mousewheel turns to reach it are
now set according to the current acitivity level.
In general: The higher the level, the deeper the user has to zoom in and
the more mousewheel turns are needed.

Also the number of nuggets to collect, is increased each level.