note_names, fixed bug related to the C note
At level 2 and others, the C note was proposed once on the right
but 2 was proposed on the left (lower and upper C).

First there was a bug, when the upper C was proposed, you was only
hearing to this note, whatever you click on the right.

Then it was confusing thus I have left the 2 C being proposed
on the right side.