note_names: alternate solution for same-sound-bug
This commit reverts a2828a2ba35246c2c479a9d57ead58cbb815183b and
provides an alternate solution for the same-sound-bug in the note_names
activity.

We keep "show only one C-button" and just fix the same-sound-bug in
lines 363 following.

Reason:
At level 4 you don't get sounds, when clicking on the buttons.
So children don't have a chance do identify the correct solution.
Their only chance is to try one with a 50:50 chance of success. This
might be a little bit frustrating.