mining: add tutorial for mining activity
This commit adds a tutorial to the mining activity. It contains
animations, which demonstrate:
  - move cursor here
  - zoom in (by mouse and touchpad)
  - click (with mouse or touchpad)
  - zoom out (by mouse and touchpad)

Based on certain conditions, the tutorial advances to its next step.

The tutorial is only enabled for level 1.

About the touchpad:
I think, the touchpad is easier recognized as such, if the keyboard above
is easier recognized. Therefor letters have been added to the former blank
keys of the keyboard.
I hope, having choosen a german keyboard layout is ok.

A BlockingArea for the space occupied by the tutorial animations
(mouse and touchpad) has been added, to avoid the nugget being placed
behind the tutorial mouse or touchpad.