mining: add button to start tutorial
Now the tutorial does not start automatically, but has to be started
explicitly by pressing the "tux-teacher" icon. (The "tux-teacher.png"
was copied from the memory activity.)

The tutorial is still only startable at level 1.

And since we start the tutorial now by a mouse-click event, we can
provide an initial mouse position for the "move to" tutorial. :)