maze: fix Tux running beyond the door in fast-mode
In the previous implementation (from 2002), fast-mode was not
immediately disabled, when Tux reached the exit. Instead of that,
Tux kept running further along an unambiguous path, until he
reached a fork or a dead end, although he already reached the
door. => He ran through the door.

This looked very strange. Now this is fixed.