removed the use of the URW font.
In some activities, the URW font was referenced. This font is
not available on all platform so it is better not to use it.

Also it is best to use the get_font from our skin system.
This way users may customize GCompris to use specialized font in
a single location.