i18n, added comments to help translation in music
The terminology of musical notation presents a particular L10n challenge as
musical notation is essential a language of it's own that has co-evolved with
it's host language.

Literal translations of quarter note, eight note, etc. by localizers not
familiar with the language of musical notation will almost certainly be
incorrect,

refs bug: https://bugzilla.gnome.org/show_bug.cgi?id=686187