explore, piano translation fixes.
When we use destop file, we set the keys to be translated as:
_key = The english string

But intltool pick the space in front of the "The" in the po file.
But in python we get the string as "The english string" thus the
translation does not match and is not displayed.

Since I don't want to break the translations at this stage, I just
added the " " in the python side for the keys we are putting this
extra space all the time.