Change python plugin init order to have _gcompris first
To use the runit.sh script, it was needed to create a link
to gcompris python plugin. This was hard to understand why
it was not working when missing.

Now we init first the internal _gcompris python plugin
and the runit script add the directory to the python plugin
so that the link is no more needed.