simplified version of the submarine activity
The submarine activity is often seen in GCompris
as being too complex, even for adults. This new
version introduces a first level in which there
is no frigate, whales and crown to catch. The door
is opened by itself.

Also now the different control items in the schema
have tooltips to explain what they are about.