lang-activity: new activity with a large word dataset
Review a set of words. Each word is shown with a voice a text and an image.
When done you are suggested an exercice in which, given the voice,
you must find the right word.

The images and voices comes from the Art4Apps project: http://www.art4apps.org/
Based in the hard organisation work of Lionel Laské in
his online version accessible at http://laske.fr/abecedarium/