Fixed wartercycle and hydroelectric activity click event
1) various svg objects which should not be clickable were missing
pointer_events = goocanvas.EVENTS_NONE
in their construction code, causing them to be clickable and steal
events from objects which should be clickable such as the sun.
I've attached a patch fixing this.

2) For some reason the bounding box of the windfarm in hydropower.py
level 2 and 3 is way too big, causing it to steal events from other
objects, like the cloud on which the user should click to generate
wind. I've tried various changes to the svg in inkscape, but failed
to fix this. In the end I ended up just flattening the 3 objects
with the 3 different positions of the windmills to bitmaps
inside inkscape, I've left the original drawn objects in the svg
with WIND_FARM_#_bak as object names. I've also attached the
resulting (work around) svg.